save progress - broken query cert
