+++ src/canisters-official-backend/Cargo.toml
@@ -19,4 +19,4 @@ serde_json = "1"
 ic-http-certification = "3.0.2"
 lazy_static = "1.5.0"
 matchit = "0.8"
-
+serde_urlencoded = "0.7.1"
