+++ src/canisters-official-backend/src/rest/templates/mod.rs
@@ -1,3 +1,4 @@
 // src/rest/templates/mod.rs
 pub mod route;
 pub mod handler;
+pub mod certs;
